--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/benbjohnson/immutable
 
 go 1.18
+
+require golang.org/x/exp v0.0.0-20220518171630-0b5c67f07fdf // indirect