This fork converts Ben's immutable types with generics. Tests are passing. I haven't really test driven this, but tests are passing...

Notes and open points:
 * List was easier to port. No major changes
 * Map: existing hashing / comparison functions are used, but adjusted
 * Map: `exp/constraints.Ordered` is used for map keys - this was mainly for the SortedMap support, but it seemed a lot easier to use the same constraint across the board. _Note that `comparable` doesn't support greater-than/less-than, only equality_. **I'm thinking it might be best to define our own constraint here, for future-proofing as much as anything. I'll aim to implement that on the weekend**
 * Map: BREAKING CHANGE: byte slices are no longer supported as hash keys, because of the Ordered constraint. I figure a `string(b)` conversion is easy enough?
 * I don't know the best way to arrange the go mod. `github.com/benjohnson/immutable/v2` seemed appropriate because it's a breaking change.
 * I don't know if there's a nicer way to design the API with further changes - I've gone for the least resistance (there was already a fair bit of resistance, but it fell into place eventually)
   * e.g. could we be using `hash/maphash` now? I experimented with it, but tests failed
 * WARNING: I did a LOT of repetitive fiddly changes here, and I may have messed up. Honestly I've been away from Go for a year and I just wanted a challenge, so this is the result. If someone is using the original, I'd be grateful for feedback.