@laher 2022-05-19
this addresses #22