@benbjohnson 2022-05-19
@laher Thanks for having a go at this. I've been on a Rust project for the last month so I haven't had a chance to dig into the Go generics yet. I took a quick once over and I think it's looking good. I'll probably defer merging this until Go 1.19 is out just so folks have time to move over. The `immutable` library is still pre-1.0 so I'm not worried about breaking compatibility. I'll just bump this to v0.4.0 once it's merged. 

In the meantime, if anyone else gets a chance to try this branch out, please drop your feedback here in this issue.