@laher 2022-05-20
Cool, makes sense. That also gives me time to ask a few best practice
questions on gophers slack #generics channel.
For now I have fixed the github action (it was on go1.15 and didn't
recognise the [T] syntax), and I removed that v2 module change.

On Fri, May 20, 2022 at 12:00 PM Ben Johnson ***@***.***>
wrote:

> @laher <https://github.com/laher> Thanks for having a go at this. I've
> been on a Rust project for the last month so I haven't had a chance to dig
> into the Go generics yet. I took a quick once over and I think it's looking
> good. I'll probably defer merging this until Go 1.19 is out just so folks
> have time to move over. The immutable library is still pre-1.0 so I'm not
> worried about breaking compatibility. I'll just bump this to v0.4.0 once
> it's merged.
>
> In the meantime, if anyone else gets a chance to try this branch out,
> please drop your feedback here in this issue.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/benbjohnson/immutable/pull/23#issuecomment-1132309803>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAA6XFIYQF5N2YOM2UHZQ2TVK3IYTANCNFSM5WL5NLFQ>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>