@benbjohnson 2022-10-02
@laher I actually still haven't used generics yet since they were released. I'll take a look over this PR tomorrow and probably merge it in. 👍