@benbjohnson 2022-10-04
I gave this a review and it looks pretty straightforward. I'll go ahead and merge and cut a new release. If anyone comes across issues with the API then we can address that in future PRs. Thanks, @laher!